requirements.txt
pyTelegramBotAPI==4.6.0
apscheduler==3.9.1
python-decouple==3.6

pytz~=2022.1
telebot~=0.0.4
psycopg2~=2.9.3